feat : 엔티티 타입 변경, 엔티티 매핑

+++ src/main/java/project/shoppingmall/domain/entity/Cart.java
@@ -14,6 +14,6 @@ public class Cart {
     private Long cart_id;
 
     @Column(nullable = false)
-    private Integer total_price;
+    private Long total_price;
 
 }
